Restructured and Refactored project structure
