+++ alembic/env.py
@@ -6,9 +6,7 @@ from sqlalchemy import pool
 
 from alembic import context
 
-from app.database import Base
-
-from app.models import *
+from app.database.database import Base
 
 from dotenv import load_dotenv
