+++ app/controllers/user_types_crud.py
@@ -1,11 +1,12 @@
 from sqlalchemy.orm import Session
-from app import models, schemas
+from app.database.models import user_types as models
+from app.database.schemas import user_types as schemas
 
 
 def get_user_type_by_id(db: Session, user_type_id: int) -> models.UserType:
     """
-    This function queries the database for the UserType with the given user_type_id
-    :param db: The database session
+    This function queries the models for the UserType with the given user_type_id
+    :param db: The models session
     :param user_type_id: The user_type_id
     :return: The UserType with the given user_type_id
     """
@@ -14,8 +15,8 @@ def get_user_type_by_id(db: Session, user_type_id: int) -> models.UserType:
 
 def get_user_type_by_name(db: Session, name: str) -> models.UserType:
     """
-    This function queries the database for the UserType with the given name
-    :param db: The database session
+    This function queries the models for the UserType with the given name
+    :param db: The models session
     :param name: The user_type_name
     :return: The UserType with the given name
     """
@@ -24,9 +25,9 @@ def get_user_type_by_name(db: Session, name: str) -> models.UserType:
 
 def get_user_types(db: Session, skip: int = 0, limit: int = 10) -> list[models.UserType]:
     """
-    This function queries the database for the UserType with the given skip and limit boundaries
+    This function queries the models for the UserType with the given skip and limit boundaries
     and returns a list of UserTypes
-    :param db: The database session
+    :param db: The models session
     :param skip: Starting index of the list, 0 by default
     :param limit: Ending index (skip + limit), 10 by default
     :return: List[Type[models.UserType]]:
@@ -37,7 +38,7 @@ def get_user_types(db: Session, skip: int = 0, limit: int = 10) -> list[models.U
 def create_user_type(db: Session, user_type: schemas.UserTypeCreate) -> models.UserType:
     """
     This function creates a new UserType
-    :param db: The database session
+    :param db: The models session
     :param user_type: schemas.UserTypeCreate
     :return: The created UserType
     """
@@ -53,7 +54,7 @@ def create_user_type(db: Session, user_type: schemas.UserTypeCreate) -> models.U
 def delete_user_type(db: Session, user_type_id: int) -> models.UserType:
     """
     This function deletes a User Type based on the given user_type_id
-    :param db: The database session
+    :param db: The models session
     :param user_type_id: The user_type_id
     :return: Deleted UserType
     """
